Move-install-script-to-image-modifier (#25)
Copy the install.sh script to here and redirect the image installation
scripts to use this version. Clean up the redundancies in the install
scripts.

Once this is accepted, the https://git.io/JJrEP short URL should be
updated to point to install.sh in this repo and it should be deleted
from the Photonvision repo.