Allow users to install any release via install.sh (#49)
This PR fixes several problems with the `install.sh` script:

1. The script relied on the GitHub rest API to get a list of releases.
The return from the API is paginated and only returns 30 items by
default. The script now has an optional `count` argument for the `-l`
option that allows the user to specify how many releases should be
returned.
2. More significantly, the script was using the returned list of
releases to determine if a release is available. Since the list was
truncated to 30 items, older releases such as the 2023.4.2 version were
not seen as available. The script now checks directly with the REST API
to determine if a release is available, so the pagination is no longer
an issue.
3. Parsing of options with optional arguments was not being done
correctly when the short version of the option was used. This is now
working correctly.
4. Added a "Test" option that skips all commands that directly make
changes to the system. This makes it easier to debug the script without
risking making unwanted changes.
5. The script used a pretty restrictive filter for which releases to
list and it wouldn't allow users to install "Dev" or "rc" releases. The
filter is now much less restrictive.
6. The script now installs the correct jammy repository for Ubuntu 24.04
on x86 systems.

NOTE: Raspberry Pi images include `libcamera-dev/stable
0.4.0+rpt20250127-3`. They will only work for PhotonVision jars that
include libcameraDriverVersion = v2025.0.4.

Documentation update:
https://github.com/PhotonVision/photonvision/pull/1746